Apply review move `FiveMinutesUser` to `testutil` package.
